Add python<3.4 compatibility
