+++ ffmpeg/_run.py
@@ -7,7 +7,6 @@ import collections
 import copy
 import operator
 import subprocess
-import asyncio
 
 from ._ffmpeg import input, output
 from .nodes import (
@@ -327,43 +326,12 @@ def run(
     return out, err
 
 
-@output_operator()
-async def run_asyncio(
-    stream_spec,
-    cmd='ffmpeg',
-    pipe_stdin=False,
-    pipe_stdout=False,
-    pipe_stderr=False,
-    quiet=False,
-    overwrite_output=False,
-):
-    """Asynchronously invoke ffmpeg in asyncio sync/await style and return coroutine.
-    Have the same possibilities as `run_async` call.
+__all__ = ['compile', 'Error', 'get_args', 'run', 'run_async']
 
-    Args:
-        pipe_stdin: if True, connect pipe to subprocess stdin (to be
-            used with ``pipe:`` ffmpeg inputs).
-        pipe_stdout: if True, connect pipe to subprocess stdout (to be
-            used with ``pipe:`` ffmpeg outputs).
-        pipe_stderr: if True, connect pipe to subprocess stderr.
-        quiet: shorthand for setting ``capture_stdout`` and
-            ``capture_stderr``.
-
-    Returns:
-        A Process instance as a coroutine
-    """
-
-    args = compile(stream_spec, cmd, overwrite_output=overwrite_output)
-    stdin_stream = asyncio.subprocess.PIPE if pipe_stdin else None
-    stdout_stream = asyncio.subprocess.PIPE if pipe_stdout or quiet else None
-    stderr_stream = asyncio.subprocess.PIPE if pipe_stderr or quiet else None
-
-    return await asyncio.create_subprocess_exec(
-        *args,
-        stdin=stdin_stream,
-        stdout=stdout_stream,
-        stderr=stderr_stream
-    )
-
-
-__all__ = ['compile', 'Error', 'get_args', 'run', 'run_async', 'run_asyncio']
+try:
+    from ._run_asyncio import *
+except (SyntaxError, ImportError):
+    pass
+else:
+    from . import _run_asyncio
+    __all__ += _run_asyncio.__all__
