+++ ffmpeg/tests/test_ffmpeg.py
@@ -10,7 +10,6 @@ import random
 import re
 import subprocess
 import sys
-import asyncio
 
 
 try:
@@ -786,29 +785,5 @@ def test__multi_output_edge_label_order():
     assert out2 == get_filter_complex_input(flt_cmpl, 'hflip')
 
 
-def test_run_asyncio():
-    async def test_async():
-        process = await (
-            ffmpeg
-             .input(TEST_INPUT_FILE1)
-             .output('pipe:', format='rawvideo', pix_fmt='rgb24')['v']
-             .run_asyncio(pipe_stdout=True, quiet=False)
-        )
-
-        video_frame_size = 320 * 240 * 3  # Note: RGB24 == 3 bytes per pixel. 320x240 - video size
-
-        total_bytes = 0
-
-        while True:
-            frame_bytes = await process.stdout.read(video_frame_size)
-            if len(frame_bytes) == 0:
-                break
-            else:
-                total_bytes += len(frame_bytes)
-
-        await process.wait()
-
-        assert total_bytes == 48153600, 'Incorrect size of the output frames'
-
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(test_async())
+if sys.version_info >= (3, 4):
+    from ._test_asyncio import test_run_asyncio
